trust-store: move the AppArmor resolution to the Skeleton class
The resolution of the AppArmor label should ideally be done by the D-Bus
daemon itself, and we should just call its GetConnectionCredentials
method. Unfortunately, dbus-cpp does not support that.

Still, move this code do the D-Bus adaptor (that is, the Skeleton class)
which is where the AppArmor label would be available if we used the
proper D-Bus method to resolve it.