data: make trust-stored service a user unit
With a reference to `mir_socket_trusted`, it's pretty clear that this
unit is intended to run as a user unit, not a system unit. Thus, update
as such.

Unfortunately, this means we now __really__ have no way to track the
restart of the main service. Perhaps it's not needed, but if it is, I
think the trust-stored-skeleton itself should be more resilient instead
of putting pressure to service manager.